requirements.txt
Flask>=2.3.0
psycopg2-binary>=2.9.0
